refactor: simplify app settings navigation
