+++ src/components/AppSettingsDialog.tsx
@@ -1,39 +1,13 @@
 import { useState } from "react";
-import { X, Keyboard, Monitor } from "lucide-react";
+import { Info, Settings, UserCog, X } from "lucide-react";
 import type { ThemeMode } from "../types";
 import { useI18n } from "../i18n";
 import s from "../styles";
-import claudeLogo from "../assets/claude.svg";
-import chatgptLogo from "../assets/chatgpt.svg";
-import appLogo from "../assets/app-logo.png";
 import { AboutPanel } from "./app-settings/AboutPanel";
-import { AgentConfigPanel } from "./app-settings/AgentConfigPanel";
-import { GeneralPanel } from "./app-settings/GeneralPanel";
-import { ShortcutsPanel } from "./app-settings/ShortcutsPanel";
-import { ThemePanel } from "./app-settings/ThemePanel";
-import { getAgentSettingsFilePath } from "./app-settings/shared";
-import type { AgentKey, AppSettingsNavItem, NavKey } from "./app-settings/types";
-
-const NAV_ITEMS: AppSettingsNavItem[] = [
-  { key: "general", labelKey: "appSettings.general" },
-  { key: "theme", labelKey: "appSettings.theme" },
-  { key: "shortcuts", labelKey: "appSettings.shortcuts" },
-  { key: "about", labelKey: "appSettings.about", logo: appLogo },
-  {
-    key: "claude",
-    labelKey: "Claude Code",
-    logo: claudeLogo,
-    filePath: getAgentSettingsFilePath("claude"),
-    lang: "json",
-  },
-  {
-    key: "codex",
-    labelKey: "Codex",
-    logo: chatgptLogo,
-    filePath: getAgentSettingsFilePath("codex"),
-    lang: "toml",
-  },
-];
+import { AgentSettingsPanel } from "./app-settings/AgentSettingsPanel";
+import { ApplicationSettingsPanel } from "./app-settings/ApplicationSettingsPanel";
+import { APP_SETTINGS_NAV_ITEMS } from "./app-settings/navigation";
+import type { NavKey } from "./app-settings/types";
 
 export function AppSettingsDialog({
   onClose,
@@ -49,21 +23,31 @@ export function AppSettingsDialog({
   onThemeModeChange: (mode: ThemeMode) => void;
 }) {
   const { t } = useI18n();
-  const [activeNav, setActiveNav] = useState<NavKey>("general");
+  const [activeNav, setActiveNav] = useState<NavKey>("app");
 
   function handleOverlayClick(e: React.MouseEvent) {
     if (e.target === e.currentTarget) onClose();
   }
 
-  const activeItem = NAV_ITEMS.find((n) => n.key === activeNav)!;
+  const activeItem = APP_SETTINGS_NAV_ITEMS.find((n) => n.key === activeNav)!;
   const activeLabel = t(activeItem.labelKey);
+
+  function renderNavIcon(key: NavKey, size: number, color?: string) {
+    if (key === "agents") {
+      return <UserCog size={size} strokeWidth={1.8} color={color} />;
+    }
+    if (key === "about") {
+      return <Info size={size} strokeWidth={1.8} color={color} />;
+    }
+    return <Settings size={size} strokeWidth={1.8} color={color} />;
+  }
 
   return (
     <div style={s.modalOverlay} onClick={handleOverlayClick}>
       <div style={s.modalBox}>
         <div style={s.settingsNav}>
           <div style={s.settingsNavTitle}>{t("appSettings.title")}</div>
-          {NAV_ITEMS.map((item) => (
+          {APP_SETTINGS_NAV_ITEMS.map((item) => (
             <button
               key={item.key}
               style={{
@@ -74,29 +58,7 @@ export function AppSettingsDialog({
               }}
               onClick={() => setActiveNav(item.key)}
             >
-              {item.logo ? (
-                <img
-                  src={item.logo}
-                  style={{ width: 14, height: 14, opacity: item.key === "codex" ? 0.7 : 1 }}
-                />
-              ) : item.key === "theme" ? (
-                <Monitor size={14} strokeWidth={1.8} />
-              ) : item.key === "shortcuts" ? (
-                <Keyboard size={14} strokeWidth={1.8} />
-              ) : (
-                <span
-                  style={{
-                    width: 14,
-                    height: 14,
-                    display: "inline-flex",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    fontSize: 13,
-                  }}
-                >
-                  ⚙
-                </span>
-              )}
+              {renderNavIcon(item.key, 14)}
               {t(item.labelKey)}
             </button>
           ))}
@@ -105,18 +67,7 @@ export function AppSettingsDialog({
         <div style={s.settingsContent}>
           <div style={s.settingsContentHeader}>
             <div style={{ display: "flex", alignItems: "center", gap: 8 }}>
-              {activeItem.logo ? (
-                <img
-                  src={activeItem.logo}
-                  style={{ width: 16, height: 16, opacity: activeItem.key === "codex" ? 0.7 : 1 }}
-                />
-              ) : activeItem.key === "theme" ? (
-                <Monitor size={16} strokeWidth={1.8} color="var(--text-secondary)" />
-              ) : activeItem.key === "shortcuts" ? (
-                <Keyboard size={16} strokeWidth={1.8} color="var(--text-secondary)" />
-              ) : (
-                <span style={{ fontSize: 15 }}>⚙</span>
-              )}
+              {renderNavIcon(activeItem.key, 16, "var(--text-secondary)")}
               <span style={s.settingsContentTitle}>{activeLabel}</span>
             </div>
             <button style={s.modalCloseBtn} onClick={onClose} title={t("common.close")}>
@@ -124,28 +75,18 @@ export function AppSettingsDialog({
             </button>
           </div>
 
-          {activeNav === "general" ? (
-            <GeneralPanel key="general" />
-          ) : activeNav === "theme" ? (
-            <ThemePanel
-              key="theme"
+          {activeNav === "app" ? (
+            <ApplicationSettingsPanel
+              key="app"
               themeMode={themeMode}
               systemPrefersDark={systemPrefersDark}
               onThemeModeChange={onThemeModeChange}
             />
-          ) : activeNav === "shortcuts" ? (
-            <ShortcutsPanel key="shortcuts" />
+          ) : activeNav === "agents" ? (
+            <AgentSettingsPanel key="agents" isDark={isDark} />
           ) : activeNav === "about" ? (
             <AboutPanel key="about" />
-          ) : (
-            <AgentConfigPanel
-              key={activeNav}
-              agentKey={activeNav as AgentKey}
-              filePath={activeItem.filePath!}
-              lang={activeItem.lang!}
-              isDark={isDark}
-            />
-          )}
+          ) : null}
         </div>
       </div>
     </div>
