+++ src/i18n.tsx
@@ -18,6 +18,7 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.general": "General",
     "appSettings.theme": "Theme",
     "appSettings.shortcuts": "Shortcuts",
+    "appSettings.agentSettings": "Agent Settings",
     "appSettings.about": "About",
     "appSettings.language": "Language",
     "appSettings.appLanguage": "App Language",
@@ -292,6 +293,7 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.general": "通用",
     "appSettings.theme": "主题",
     "appSettings.shortcuts": "快捷键",
+    "appSettings.agentSettings": "Agent 设置",
     "appSettings.about": "关于",
     "appSettings.language": "语言",
     "appSettings.appLanguage": "应用语言",
